% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bacteriaDetection.R
\name{bacteriaDetection}
\alias{bacteriaDetection}
\alias{bacteriaDetection_Single}
\alias{bacteriaDetection}
\title{Automatic bacteria detection.}
\usage{
bacteriaDetection_Single(originalImageFileName)

bacteriaDetection(originalImageFileNameList)
}
\arguments{
\item{originalImageFileName}{The input image file name. Should be provided as a combination of letters and numbers, since letters are interpreted as image source labels and numbers as image IDs.}

\item{originalImageFileNameList}{A list of image file names.}
}
\description{
Automatic bacteria detection.
}
