% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bacteriaClassification.R
\name{bacteriaClassification}
\alias{bacteriaClassification}
\title{Classification of bacteria based on their morphological features.}
\usage{
bacteriaClassification(trainDF, testDF, validDF = NULL, seed = 12345,
  outputDir = "./Results/", max_mem_size = "6G")
}
\arguments{
\item{trainDF}{A training dataframe.}

\item{testDF}{A testing dataframe.}

\item{validDF}{(Optional) A dataframe for external validation.}

\item{seed}{A random seed.}

\item{outputDir}{A directory for saving outputs.}

\item{max_mem_size}{A memory limit for H2O Java machine learning engine.}
}
\description{
Classification of bacteria based on their morphological features.
}
